+++ nbproject/private/private.xml
@@ -2,12 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/home/testo/LibElecFront/src/java/model/VerificateurSaisie.java</file>
-            <file>file:/home/testo/LibElecFront/web/resources/css/style.css</file>
-            <file>file:/home/testo/LibElecFront/web/WEB-INF/espaceClient/pageConnexion.jsp</file>
-            <file>file:/home/testo/LibElecFront/src/java/controller/FrontController.java</file>
-            <file>file:/home/testo/LibElecFront/web/WEB-INF/espaceClient/pageInscription.jsp</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
