--- src/java/UtilDB/DAOInterface.java
@@ -1,9 +1,0 @@
-package UtilDB;
-
-
-public interface DAOInterface {
-    public void insert();
-    public void update();
-    
-    
-}
