+++ src/java/accessDB/DAOInterface.java
@@ -1,0 +1,9 @@
+package accessDB;
+
+
+public interface DAOInterface {
+    public void insert();
+    public void update();
+    
+    
+}
