+++ src/java/controller/FrontController.java
@@ -19,7 +19,7 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
-import modelTablesDB.ClientDAO;
+import accessDB.ClientDAO;
 import model.VerificateurSaisie;
 
 //TODO il faut activer la session des la rentree du mot de passe
@@ -107,12 +107,21 @@ public class FrontController extends HttpServlet {
             page ="/WEB-INF/espaceClient/pageInscription.jsp";
         }
         
+        
+        //TODO il faut modifier les infos clients, 
+        //TODO il faut supprimer un compte client (changer attribut statut dans la base sans supprimer)
+        
         if("enregisterNouveauMembre".equalsIgnoreCase(section)) {
             ArrayList<String> listeSaisie = new ArrayList(15);
             HashMap<String, String> listeChaineInscription;
             
             LocalDate toDay = LocalDate.now();
             
+            /**
+             * recuperation toute saisie utilisateur
+             * rajout l'ensemble de ces saisies à une arraylist 
+             * pour permettre de les checker
+             */
             
             String genre = request.getParameter("genre");
             listeSaisie.add(genre);
@@ -124,8 +133,8 @@ public class FrontController extends HttpServlet {
             String prenom = request.getParameter("prenom");
             listeSaisie.add(prenom);
             
-             System.out.println("listeSaisie " + request.getParameter("prenom") + " / " +prenom);
-           
+            System.out.println("listeSaisie " + request.getParameter("prenom") + " / " +prenom);
+            
             
             String email = request.getParameter("email");
             listeSaisie.add(email);
@@ -142,24 +151,33 @@ public class FrontController extends HttpServlet {
             
             
             verificateurSaisie = new VerificateurSaisie();
-            
-            
             Client nouveauMembre;
+            
+            
+            /**
+            * choix d'une hashmap pour pouvoir reperer 
+            * et récupérer facilement une valeur correspondant à une clé
+            */
             HashMap<String, String> infosNouveau = new HashMap(15);
+                        
+            infosNouveau.put("genre", request.getParameter("genre"));
+            infosNouveau.put("nom", request.getParameter("nom"));
+            infosNouveau.put("prenom", request.getParameter("prenom"));
+            infosNouveau.put("email", request.getParameter("email"));
+            infosNouveau.put("password", request.getParameter("password"));
+            infosNouveau.put("telF", request.getParameter("telF"));
             
-            
-                infosNouveau.put("genre", request.getParameter("genre"));
-                infosNouveau.put("nom", request.getParameter("nom"));
-                infosNouveau.put("prenom", request.getParameter("prenom"));
-                infosNouveau.put("email", request.getParameter("email"));
-                infosNouveau.put("password", request.getParameter("password"));
-                infosNouveau.put("telF", request.getParameter("telF"));
-             
-            
+            /**
+             * On envoie la hashmap en parametre au l'objet verificateurSaisie
+             * cela retourne une autre hashmap correspondant à l'index saisie (ex: nom)
+             * et sa valeur contenant 'le msg d'erreur' ou 'ok' si la saisie est correcte
+             * Si l'inscription est invalide : la hashmap va permettre d'affecter les valeurs
+             * et reperer les saisie concernees dans la jsp 'pageInscription'
+             */
             listeChaineInscription = verificateurSaisie.checkSaisieNouveauMembre(infosNouveau);
             
+            /* drapeau permettant de savoir si l'inscription est complete ou correcte */
             nouveauComplet = verificateurSaisie.getNouveauComplet();
-            System.out.println("flag : " + nouveauComplet);
             
             if(nouveauComplet) {
                 nouveauMembre = new Client();
@@ -172,10 +190,8 @@ public class FrontController extends HttpServlet {
                         break;
                 }
                 
-                
                 nouveauMembre.setCliNom(infosNouveau.get("nom"));
                 nouveauMembre.setCliPrenom(infosNouveau.get("prenom"));
-               
                 nouveauMembre.setCliDateAdhesion(toDay.toString());
                 nouveauMembre.setCliEmail(infosNouveau.get("email"));
                 nouveauMembre.setCliMdp(infosNouveau.get("password"));
@@ -184,37 +200,46 @@ public class FrontController extends HttpServlet {
                 
                 nouveauMembre.setCliStatut(1);
                 
+                
+                
                 ClientDAO clientDao = new ClientDAO(nouveauMembre);
                 clientDao.insert();
-            }
-           
                 
-                listeChaineInscription.forEach( (k, v) -> {
-                    String tmp = (String) v;
-                    
-                    if(tmp.equals("choix obligatoire")) {
-                        request.setAttribute((String) k, " * choix obligatoire");
-                        request.setAttribute("chaineInscriptionInvalide", chaineInscriptionInvalide);
-                    }
-                    else if(tmp.equals("vide")) {
-                        request.setAttribute((String) k, " * vide");
-                        request.setAttribute("chaineInscriptionInvalide", chaineInscriptionInvalide);
-                    }
-                    else if(tmp.equals("invalide")) {
-                        request.setAttribute((String) k, " * invalide");
-                        request.setAttribute("chaineInscriptionInvalide", chaineInscriptionInvalide);
-                    }
-                    else if(tmp.equals("password")) {
-                        request.setAttribute((String) k, " * mot de passe invalide");
-                        System.out.println("k : " + k);
-                        request.setAttribute("chaineInscriptionInvalide", chaineInscriptionInvalide);
-                    }
-                });
+                // TODO il faut controler si le nouveau membre existe deja dans la base 
+                request.setAttribute("client", nouveauMembre);
+            }
             
-            
-            
+            listeChaineInscription.forEach( (k, v) -> {
+                String tmp = (String) v;
+                
+                if(tmp.equals("choix obligatoire")) {
+                    request.setAttribute((String) k, " * choix obligatoire");
+                    request.setAttribute("chaineInscriptionInvalide", chaineInscriptionInvalide);
+                }
+                else if(tmp.equals("vide")) {
+                    request.setAttribute((String) k, " * vide");
+                    request.setAttribute("chaineInscriptionInvalide", chaineInscriptionInvalide);
+                }
+                else if(tmp.equals("invalide")) {
+                    request.setAttribute((String) k, " * invalide");
+                    request.setAttribute("chaineInscriptionInvalide", chaineInscriptionInvalide);
+                }
+                else if(tmp.equals("password")) {
+                    request.setAttribute((String) k, " * mot de passe invalide");
+                    System.out.println("k : " + k);
+                    request.setAttribute("chaineInscriptionInvalide", chaineInscriptionInvalide);
+                }
+            });
+                
+            if(nouveauComplet) {
+                
+               
+                page = "/WEB-INF/espaceClient/espacePersonnel.jsp";
+                
+            }
+            else {
             page = "/WEB-INF/espaceClient/pageInscription.jsp";
-            
+            }
             
             //}
 //adresse
